Added Blockout and a Player Reference I made to aid in scaling.
Along with the blockout, I took the liberty of cleaning up the default scene so that it's only the floor, outer edges and some of the miscellaneous stuff I figured we'd want to keep! If I've messed up something, however, we can revert to the previous backup and I can import only the blockout!